refactor ascend ci diff scan skill code
