+++ .agent/skills/ascend-ci-case-diff-scan/SKILL.md
@@ -54,7 +54,6 @@ Recognize only workflow commands that visibly execute tests:
 - `pytest ... tests/...`
 - `bash tests/.../*.sh`
 - `torchrun ... tests/...`
-- `torchrun ... -m pytest tests/...`
 
 For each extracted case, preserve:
