+++ .agent/skills/ascend-ci-case-diff-scan/scripts/modules/render.py
@@ -1,0 +1,132 @@
+#!/usr/bin/env python3
+# Copyright 2025 Bytedance Ltd. and/or its affiliates
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Markdown report rendering for the Ascend CI case diff scanner."""
+
+from __future__ import annotations
+
+from itertools import zip_longest
+
+
+def render_reference_details(indent: str, label: str, ref: dict | None) -> list[str]:
+    """Render one workflow location with enough context for manual navigation."""
+    lines = [f"{indent}- {label}:"]
+    if not ref:
+        lines.append(f"{indent}  - None")
+        return lines
+    lines.append(f"{indent}  - Workflow file: `{ref['workflow_path']}`")
+    lines.append(f"{indent}  - Line number: `{ref['line_number']}`")
+    lines.append(f"{indent}  - Workflow context: `{ref['workflow_name']} / {ref['job_name']} / {ref['step_name']}`")
+    lines.append(f"{indent}  - Command: `{ref['raw_command']}`")
+    return lines
+
+
+def render_adjacent_pairs(cpu_gpu_refs: list[dict], npu_refs: list[dict]) -> list[str]:
+    """Render CPU/GPU and NPU references next to each other."""
+    lines: list[str] = []
+    if not cpu_gpu_refs and not npu_refs:
+        lines.append("  - None")
+        return lines
+    sorted_cpu_gpu = sorted(cpu_gpu_refs, key=lambda item: (item["name"], item["workflow_path"], item["line_number"]))
+    sorted_npu = sorted(npu_refs, key=lambda item: (item["name"], item["workflow_path"], item["line_number"]))
+    for index, (cpu_gpu_ref, npu_ref) in enumerate(zip_longest(sorted_cpu_gpu, sorted_npu), start=1):
+        lines.append(f"  - Pair {index}")
+        lines.extend(render_reference_details("    ", "CPU/GPU", cpu_gpu_ref))
+        lines.extend(render_reference_details("    ", "NPU", npu_ref))
+    return lines
+
+
+def render_table(headers: list[str], rows: list[list[str]]) -> list[str]:
+    """Render a Markdown table."""
+    lines = [
+        "| " + " | ".join(headers) + " |",
+        "| " + " | ".join(["---"] * len(headers)) + " |",
+    ]
+    for row in rows:
+        lines.append("| " + " | ".join(row) + " |")
+    return lines
+
+
+def render_case_section(title: str, items: list[dict]) -> list[str]:
+    """Render one details subsection."""
+    lines = [f"### {title}", ""]
+    if not items:
+        lines.append("- None")
+        lines.append("")
+        return lines
+    for item in items:
+        lines.append(f"- Case path: `{item['name']}`")
+        lines.extend(render_adjacent_pairs(item["cpu_gpu_refs"], item["npu_refs"]))
+    lines.append("")
+    return lines
+
+
+def render_details_block(title: str, comparison: dict[str, list[dict]]) -> list[str]:
+    """Render UT or ST comparison details in the requested order."""
+    lines = [f"## {title}", ""]
+    lines.extend(render_case_section("Matched Cases", comparison["matched"]))
+    lines.extend(render_case_section("CPU/GPU-only Cases", comparison["cpu_gpu_only"]))
+    lines.extend(render_case_section("NPU-only Cases", comparison["npu_only"]))
+    lines.extend(render_case_section("Manual Review", comparison["manual_review"]))
+    return lines
+
+
+def render_report(report: dict) -> str:
+    """Render the final Markdown report."""
+    lines = [
+        "# Ascend CI Workflow and Case Alignment Report",
+        "",
+        "Read this report in order: Ignored Workflows, Scanned Workflows, UT Case Details, and ST Case Details.",
+        "",
+        (
+            "Within both UT Case Details and ST Case Details, cases are grouped into Matched Cases, "
+            "CPU/GPU-only Cases, NPU-only Cases, and Manual Review."
+        ),
+        "",
+        f"## Ignored Workflows ({len(report['ignored_workflows'])})",
+        "",
+    ]
+    ignored_rows = [[path] for path in report["ignored_workflows"]]
+    if ignored_rows:
+        lines.extend(render_table(["Workflow Name"], ignored_rows))
+    else:
+        lines.append("No workflows were ignored.")
+    lines.extend(
+        [
+            "",
+            "## Scanned Workflows",
+            "",
+        ]
+    )
+    scanned_rows = [
+        [
+            row["workflow_name"],
+            str(row["cpu_gpu_case_count"]),
+            str(row["npu_supported_case_count"]),
+        ]
+        for row in report["scanned_workflows"]
+    ]
+    if scanned_rows:
+        lines.extend(
+            render_table(
+                ["Workflow Name", "CPU/GPU Case Count", "NPU Supported Case Count"],
+                scanned_rows,
+            )
+        )
+    else:
+        lines.append("No workflows were scanned.")
+    lines.extend(["", *render_details_block("UT Case Details", report["ut_details"])])
+    lines.extend(["", *render_details_block("ST Case Details", report["st_details"])])
+    return "\n".join(lines).rstrip() + "\n"
